+++ src/ui/command/Command.h
@@ -5,7 +5,7 @@
 #include "Result.h"
 #include "Error.h"
 #include "ui/controller/Controller.h"
-#include "model/TasksSortBy.h"
+#include "utilities/TasksSortBy.h"
 
 namespace ui::command {
